Added Velocity updates Numerical calculations definitely wrong Features: - Found the appropriate spot in physics update to change PhysicsVelocity - Added compile-time gravity constant (still non-physical units) - Velocity is updating (not sure about correctness, but its clearly updating)
Bugs:
- Gravity and pressure appear to both drive outward push
- Reversing gravity leads to resource issues with hashtable
- Kernel derivative calculations are likely an issue here.

Next steps are likely to alter the storage of Kernel calculations
and interacting pairs. The current hashmap scheme for holding particle
pairs and kernels means we would have to resize hash-maps during a
multithreaded parallel job. This resize is a nonstarter using the
current unity containers. Unity.Physics uses sorted Arrays and lists
to process this type of data multithreaded. We should do the same.

There appears to be a way to hold interaction pairs and kernel parameters
via DynamicArray properties that leverages the existing component
infrastructure. Not sure if that will work or if we will have to write
a separate NativeList/NativeArray.