Unity Physics 0.6.0-preview.3 as a local package Starting point for enabling iteration over internal physics data structures
I plan to use physics data structures for two major things:
- O(NlogN) Gravity
- pre-calculation of interaction kernels for Density/Pressure calculations

Other ways to do this which I have rejected (for now):

Collision Queries using Physics standard interface:
I have not prototyped this, but it seems like it would have
similar performance to the chosen method, while having an implementation that
repeats work. If we have N particles and average M interactions per particle:
We will take O(N log N) time doing collision queries against the BVH across N
particles in order to obtain an O(N*M) long interaction list. At the moment
we would do this work, very-nearly the interaction list we want already sits
in memory, having been assembled during broadphase, and we need only filter
it, which will be closer to O(N*M), especially as N grows and M remains a
constant for larger simulations.

Reflection to gain access to types:
This is possible, but requires significant overhead the more types that need
to be unpacked. Physics uses several nested internal types for scheduling
information. There is significant work here in simply unpacking the types,
and because we access those types in multithreaded code, we many not be able
to achieve performance without some caching scheme for the relevant memory
offsets from the types. This is a lot more work than changing keywords in
physics, and letting the compiler do the memory math.